Move toComponentFieldViewState to ComponentFieldDelegateState
COAND-935
