+++ components-core/src/main/java/com/adyen/checkout/components/core/internal/ui/model/ComponentFieldDelegateState.kt
@@ -28,5 +28,20 @@ fun <T> ComponentFieldDelegateState<T>.updateFieldState(
     value = value ?: this.value,
     validation = validation ?: this.validation,
     hasFocus = hasFocus ?: this.hasFocus,
-    shouldHighlightValidationError = shouldHighlightValidationError ?: this.shouldHighlightValidationError,
+    shouldHighlightValidationError = shouldHighlightValidationError
+        ?: this.shouldHighlightValidationError,
 )
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+fun <T> ComponentFieldDelegateState<T>.toComponentFieldViewState() =
+    ComponentFieldViewState(
+        value = value,
+        hasFocus = hasFocus,
+        errorMessageId = takeIf { fieldState ->
+            fieldState.shouldShowValidationError()
+        }?.validation.let { it as? Validation.Invalid }?.reason,
+    )
+
+// Validation error should be shown, when the field loses its focus or when we manually trigger a validation
+internal fun <T> ComponentFieldDelegateState<T>.shouldShowValidationError() =
+    !this.hasFocus || this.shouldHighlightValidationError
