+++ mbway/src/main/java/com/adyen/checkout/mbway/internal/ui/model/MBWayViewState.kt
@@ -8,9 +8,8 @@
 
 package com.adyen.checkout.mbway.internal.ui.model
 
-import com.adyen.checkout.components.core.internal.ui.model.ComponentFieldDelegateState
 import com.adyen.checkout.components.core.internal.ui.model.ComponentFieldViewState
-import com.adyen.checkout.components.core.internal.ui.model.Validation
+import com.adyen.checkout.components.core.internal.ui.model.toComponentFieldViewState
 import com.adyen.checkout.ui.core.internal.ui.model.CountryModel
 
 internal data class MBWayViewState(
@@ -24,16 +23,3 @@ internal fun MBWayDelegateState.toViewState() = MBWayViewState(
     countryCodeFieldState = this.countryCodeFieldState.toComponentFieldViewState(),
     phoneNumberFieldState = this.localPhoneNumberFieldState.toComponentFieldViewState(),
 )
-
-internal fun <T> ComponentFieldDelegateState<T>.toComponentFieldViewState() =
-    ComponentFieldViewState(
-        value = value,
-        hasFocus = hasFocus,
-        errorMessageId = takeIf { fieldState ->
-            fieldState.shouldShowValidationError()
-        }?.validation.let { it as? Validation.Invalid }?.reason,
-    )
-
-// Validation error should be shown, when the field loses its focus or when we manually trigger a validation
-internal fun <T> ComponentFieldDelegateState<T>.shouldShowValidationError() =
-    !this.hasFocus || this.shouldHighlightValidationError
